requirements.txt
python-telegram-bot==13.7
pyaml==20.4.0